user_service/requirements.txt
Flask==2.0.1
mysql-connector-python==8.0.26
cryptography==3.4.7
PyJWT==2.1.0
Werkzeug==2.0.1
Pillow==8.3.1